Remove empty rows/columns from the range returned by worksheet::rows(true)
Note that the iterator already skipped those empty rows/cells, but those
rows/columns were included in the returned range.reference(). If someone is
interested in the old range_reference (i.e. the envolope of all
non-empty cells or rows with properties), calculate_dimension(true) can
be called directly.

Note that Google Sheets sometimes add empty rows (with only properties
being defined). This results in a different range being returned by
rows/columns before or after the file is saved by Google Sheets.

Note that the (old) default behaviour of calculate_dimension is kept, as
this is how the dimensions are stored in the worksheet. Changing this
behaviour results in failed unit tests.

Fixes #5