Use XLNT_API_INTERNAL for internal API (#44)
Note that exporting is needed to be able to unit test those internal
APIs.

Moved testing of internal API to a seperate folder, to allow testing
whether the public interface is correctly exported even if the internal
API is not exported. Only tested for dynamic library to limit the number
of CI jobs (if it works for dynamic, it should work for static too)

Disabled tests while benchmarking to avoid running tests twice.

Hide symbols not explicitly marked for export on all platforms.

Removed XLNT_API from enum classes to prevent compiler warning. Note
that exporting enum class is not needed.