Fix generating coverage report on master branch (#80)
If coverage is generated for a merged PR (i.e. a new commit on the
master branch), the coverage should be compared to the previous master
commit, otherwise it's compared with itself.

"-f" option used for curl to ensure the job fails in case the request
fails.

Related to PR #79