Revert interface changes of cell_reference and range_reference
Before this PR, a cell_reference and a range_reference were guaranteed
to be valid references if no exception was thrown. This commit restores
this behaviour for backwards compatibility. Another design decision may
be made in case of a major version upgrade.

Adapted xlsx consumer to correctly handle #REF! in defined names.