Patch for dealing with file containing invalid relations
An issue was find in ticket #109 where an excel file may contain invalid relations. The relation was expressed as a relative path, but when canonicalized, it would break out, and lead to a segfault while popping an element from an empty vector. This commit will fix the parsing by ignoring vector pops if the vector is empty. This behavior can be disabled by defining THROW_ON_INVALID_XML. In that case, the an invalid_relation exception will be thrown.